+++ demos/showcase/src/main/scala/com/sfxcode/sapphire/extension/showcase/controller/AbstractBaseController.scala
@@ -1,0 +1,13 @@
+package com.sfxcode.sapphire.extension.showcase.controller
+
+import com.sfxcode.sapphire.core.controller.ViewController
+
+class AbstractBaseController extends ViewController {
+
+  def mainWindowController: MainWindowController = {
+    parent.asInstanceOf[MainWindowController]
+  }
+
+  def workspaceManager = mainWindowController.workspaceManager
+
+}