+++ src/it/scala/com/sfxcode/sapphire/extension/demo/model/BookRating.scala
@@ -1,6 +1,0 @@
-package com.sfxcode.sapphire.extension.demo.model
-
-/**
- * Created by tom on 06.07.15.
- */
-case class BookRating(id: Long, name: String, rating: Double, pages: Double)